realm: Add parameter to rotate kdcs and admin servers arrays
We have observed that clients prefer to talk to KDCs in the order
they're listed in `/etc/krb5.conf`.  To more evenly distribute the load
across a list of servers, add a new parameter, `rotate_servers`,
disabled by default, which, when enabled, uses the `fqdn_rotate` Puppet
function to idempotently "reorder" the list of KDCs and admin servers.